+++ cpp/tests/systemintegration-tests/AsyncProxyBuilderTest.cpp
@@ -72,7 +72,8 @@ TEST_F(AsyncProxyBuilderTest, createProxyAsync_succeeds)
 
     runtime->registerProvider<tests::testProvider>(domain, testProvider, providerQos);
 
-    ProxyBuilder<tests::testProxy>* testProxyBuilder = runtime->createProxyBuilder<tests::testProxy>(domain);
+    std::unique_ptr<ProxyBuilder<tests::testProxy>> testProxyBuilder =
+            runtime->createProxyBuilder<tests::testProxy>(domain);
 
     Semaphore onSuccessCalledSemaphore;
 
@@ -95,7 +96,8 @@ TEST_F(AsyncProxyBuilderTest, createProxyAsync_succeeds)
 
 TEST_F(AsyncProxyBuilderTest, createProxyAsync_exceptionThrown)
 {
-    ProxyBuilder<tests::testProxy>* testProxyBuilder = runtime->createProxyBuilder<tests::testProxy>("unknownDomain");
+    std::unique_ptr<ProxyBuilder<tests::testProxy>> testProxyBuilder =
+            runtime->createProxyBuilder<tests::testProxy>("unknownDomain");
 
     Semaphore onErrorCalledSemaphore;