+++ cpp/tests/systemintegration-tests/End2EndSubscriptionTest.cpp
@@ -143,7 +143,7 @@ protected:
     }
 
     std::unique_ptr<tests::testProxy> buildProxy() {
-        ProxyBuilder<tests::testProxy>* testProxyBuilder
+        std::unique_ptr<ProxyBuilder<tests::testProxy>> testProxyBuilder
                 = runtime2->createProxyBuilder<tests::testProxy>(domainName);
         DiscoveryQos discoveryQos;
         discoveryQos.setArbitrationStrategy(DiscoveryQos::ArbitrationStrategy::HIGHEST_PRIORITY);
@@ -157,7 +157,6 @@ protected:
                 ->setCached(false)
                 ->setDiscoveryQos(discoveryQos)
                 ->build();
-        delete testProxyBuilder;
         return std::move(testProxy);
     }