+++ tests/performance-test/src/main/cpp/short-circuit/ShortCircuitTest.h
@@ -48,8 +48,8 @@ struct ShortCircuitTest : public PerformanceTest
         echoProviderQos.setSupportsOnChangeSubscriptions(true);
         runtime.registerProvider<tests::performance::EchoProvider>(
                 domainName, echoProvider, echoProviderQos);
-        std::unique_ptr<ProxyBuilder<tests::performance::EchoProxy>> proxyBuilder(
-                runtime.createProxyBuilder<tests::performance::EchoProxy>(domainName));
+        std::unique_ptr<ProxyBuilder<tests::performance::EchoProxy>> proxyBuilder =
+                runtime.createProxyBuilder<tests::performance::EchoProxy>(domainName);
         echoProxy = proxyBuilder->setDiscoveryQos(joynr::DiscoveryQos())->build();
     }