+++ wiki/ReleaseNotes.md
@@ -3,6 +3,7 @@
 ## API relevant changes
 
 * **[C++]** Proxy builder returns a std::unique_ptr to the created proxy instead of a raw pointer.
+* **[C++]** Joynr runtime returns a std::unique_ptr to a created proxy builder instead of a raw pointer.
 
 ## Other changes
 None.